feat: Product 생성 시 dto로 요청

+++ src/main/java/kitchenpos/domain/menu/Product.java
@@ -17,7 +17,7 @@ public class Product extends BaseEntity {
 
     }
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, Long price) {
         this.name = name;
         this.price = new Price(price);
     }
